+++ src/MainNav/index.ts
@@ -1,0 +1,3 @@
+import { MainNav, NavigationElement } from './MainNav';
+
+export { MainNav, NavigationElement };